Move FunctionExpression to expressions package
