+++ java-vtl-script/src/main/java/no/ssb/vtl/script/expressions/FunctionExpression.java
@@ -1,4 +1,4 @@
-package no.ssb.vtl.script.expression;
+package no.ssb.vtl.script.expressions;
 
 /*
  * ========================LICENSE_START=================================
@@ -9,9 +9,7 @@ package no.ssb.vtl.script.expression;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
